Refactor: Comments and Naming issues
Changed the names of some components to better reflect their functionality. Added additional comments to provide clarity
