+++ mystic_tuner_frontend/src/components/header.jsx
@@ -2,14 +2,14 @@
 import {auth0} from '@/lib/auth0.js';
 import LoginDropdown from './loginDropdown';
 import HomeButton from './homeButton';
-import CustomLine from './customLine';
+import TitleBar from './titleBar';
 
 export default async function Header(){
     const session = await auth0.getSession();
     return(
         <div className='flex justify-between'>
             <HomeButton/>
-            <CustomLine/>
+            <TitleBar/>
             <LoginDropdown session={session}/>
         </div>
     );
