+++ mystic_tuner_frontend/src/components/homeButton.jsx
@@ -12,20 +12,13 @@ export default function HomeButton(){
 
     return(
         <div className="flex justify-start items-center">
-            <Button 
-                radius="full" 
-                isIconOnly 
-                variant="bordered" 
-                className="w-[50] h-[50] border-0 p-0 m-0"
-                as={Link}
-                href="/"
-                >
-                        <Image 
-                            src={smallLogo} 
-                            width={iconWidth} 
-                            height={iconHeight} 
-                            alt='Mystic Tuner Logo'
-                        />
+            <Button radius="full" isIconOnly variant="bordered" 
+                className="w-[50] h-[50] border-0 p-0 m-0" as={Link} href="/">
+                <Image 
+                    src={smallLogo} 
+                    width={iconWidth} 
+                    height={iconHeight} 
+                    alt='Mystic Tuner Logo'/>
             </Button>
         </div>
     );
